ref: Move upstream fetching into its own actor
