+++ relay-server/src/actors/mod.rs
@@ -48,6 +48,7 @@ pub mod project;
 pub mod project_cache;
 pub mod project_keys;
 pub mod project_local_cache;
+pub mod project_upstream_cache;
 pub mod server;
 pub mod upstream;
