+++ relay-server/src/endpoints/project_configs.rs
@@ -2,7 +2,8 @@ use actix::prelude::*;
 use actix_web::{Error, Json};
 use futures::{future, Future};
 
-use crate::actors::project_cache::{GetProject, GetProjectStates, GetProjectStatesResponse};
+use crate::actors::project_cache::GetProject;
+use crate::actors::project_upstream_cache::{GetProjectStates, GetProjectStatesResponse};
 
 use crate::actors::project::GetProjectState;
 use crate::extractors::{CurrentServiceState, SignedJson};
